finish fraction tokenizer
